[Feat] Setup the basic site structure
The basic site overview is established. The docs wil be written
in markdown with different routing per page. The reactjs.org repo
will serve as a guide on how to accomplish the integration.

Note: Unfortunately I lumped a bunch of random changes in here to
that are unfinished and make the PR much larger than intended.

Summary of Changes
--Add Bootstrap Sass
Here we add bootstrap 4 support. However the library is in beta release
meaning there might be points of breakage in the future. We must ensure
that we are aware of the release cycles for this library to update the
codebase accordingly.

--Add react for bootstrap
Reactstrap is a reactjs component library for bootstrap 4.
We can correctly assume that this library is in some unstable
phase therefore there must be great care given to following
the development pattern and updating our code accordingly.

--Add the initial landing page text

The landing page as the inital version of data to display.
Next it needs coloring and all the other product pages to
be filled out.

--Add components for navbar and hero

--Add collection view for themes
The gallery page continues all the themes we will be making
accessible to the public.

--Add pages for docs and contribute