[Fix] reorganize files to explicitly import styles used by the page
The reorganization is a direct result of me wanting to modify the body
and html tag on the themes-detail page. The styles were imported through
a pipeline where order mattered. The styles compounded on eachother thus
introducing an override in one location caused the override to trickle
down into the next set of styles in a somewhat hidden manner. The new
format makes sure each component is imported with its own styles to
ensure isolated stylying.

Caveat is that now you need to be aware that global styles such as
jumbotron or navbar maybe now be imported in an order that might
override styles. To prevent this ensure that naming conventions produce
collision free styles.