[Feat] style themes detail product page
The page has the format of how the products should be presented to the
user. In the future it would be great to load the actual components into
the web page and have the user click through the demo site. For now
images will suffice.

Some product pages might now have these are nicely scrollable such as
massively. In this case we should use a muted background with several
images on the main area that are generously spaced apart.