[Feat] netflix style redesign
The previous design focused on Jay however the core focus of this site
are the themes. The new design pushes users to view themes and install
them using Jay.

The CLI tool is placed throughout the site. In the themes details page
there will be quick references to the tool either readme of the github
repo or the blog post that details how Jay is used.

The current design pattern is modeled after Netflix. Use the site as a
guide to how the text and themes should interact with the page.