[Feat] theme gallery redesign (#21)
The theme gallery features a Netflix style lineup. The cards have a title gradient with no margin. There needs to be adjustment of the title and gradient on different screen sizes. Some images conflict with HTML elements on the screen. For example the hamburger icon on Lens, the gradient on Tessellate.